Move FE sorting of Policies table to promise resolve to avoid re-render
